refactor: separation in packages
