+++ src/main/java/ulaval/glo2003/Main.java
@@ -8,8 +8,8 @@ import org.glassfish.grizzly.http.server.HttpServer;
 import org.glassfish.jersey.grizzly2.httpserver.GrizzlyHttpServerFactory;
 import org.glassfish.jersey.server.ResourceConfig;
 import ulaval.glo2003.api.HealthResource;
-import ulaval.glo2003.api.ProductResource;
-import ulaval.glo2003.api.SellerResource;
+import ulaval.glo2003.api.product.ProductResource;
+import ulaval.glo2003.api.seller.SellerResource;
 import ulaval.glo2003.api.exceptionHandling.NotFoundExceptionMapper;
 import ulaval.glo2003.api.exceptionHandling.ParamExceptionMapper;
 import ulaval.glo2003.domain.Seller;
