+++ src/main/java/ulaval/glo2003/api/product/ProductResponse.java
@@ -1,4 +1,4 @@
-package ulaval.glo2003.api;
+package ulaval.glo2003.api.product;
 
 import ulaval.glo2003.domain.Product;
