+++ src/main/java/ulaval/glo2003/api/seller/SellerRequest.java
@@ -1,4 +1,4 @@
-package ulaval.glo2003.api;
+package ulaval.glo2003.api.seller;
 
 import ulaval.glo2003.api.exceptionHandling.MissingParamException;
