+++ src/main/java/ulaval/glo2003/api/seller/SellerResource.java
@@ -1,4 +1,4 @@
-package ulaval.glo2003.api;
+package ulaval.glo2003.api.seller;
 
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.Context;
@@ -6,6 +6,7 @@ import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.UriInfo;
 import java.util.List;
+
 import ulaval.glo2003.domain.Seller;
 
 @Path("/sellers")
