+++ src/main/java/ulaval/glo2003/api/seller/SellerResponse.java
@@ -1,6 +1,8 @@
-package ulaval.glo2003.api;
+package ulaval.glo2003.api.seller;
 
 import java.util.ArrayList;
+
+import ulaval.glo2003.api.product.ProductResponse;
 import ulaval.glo2003.domain.Seller;
 
 public class SellerResponse {
